cotroller_update
